Refactor teams code to separate file. Currently only used by cycles. Split out cycle run tests.
