+++ wrfhydropy/tests/test_cycle.py
@@ -1066,7 +1066,7 @@ def test_cycle_ensemble_parallel_compose(
     assert time_taken < 1.5
 
 
-def test_cycle_run(
+def test_cycle_serial(
     simulation_compiled,
     job_restart,
     scheduler,
@@ -1075,14 +1075,11 @@ def test_cycle_run(
     init_times,
     restart_dirs
 ):
-
     sim = simulation_compiled
-
     tmp = pathlib.Path(tmpdir)
     forcing_dirs = [tmp / letter for letter in list(string.ascii_lowercase)[0:len(init_times)]]
     for dir in forcing_dirs:
         dir.mkdir()
-
     cy = CycleSimulation(
         init_times=init_times,
         restart_dirs=restart_dirs,
@@ -1090,7 +1087,6 @@ def test_cycle_run(
     )
     cy.add(sim)
     cy.add(job_restart)
-
     # Serial test
     cy_serial = copy.deepcopy(cy)
     cy_dir = pathlib.Path(tmpdir).joinpath('cycle_serial_run')
@@ -1103,6 +1099,28 @@ def test_cycle_run(
         "Some serial cycle casts did not run successfully."
     assert cy_dir.joinpath("WrfHydroCycle.pkl").exists()
 
+
+def test_cycle_run_parallel(
+    simulation_compiled,
+    job_restart,
+    scheduler,
+    tmpdir,
+    capfd,
+    init_times,
+    restart_dirs
+):
+    sim = simulation_compiled
+    tmp = pathlib.Path(tmpdir)
+    forcing_dirs = [tmp / letter for letter in list(string.ascii_lowercase)[0:len(init_times)]]
+    for dir in forcing_dirs:
+        dir.mkdir()
+    cy = CycleSimulation(
+        init_times=init_times,
+        restart_dirs=restart_dirs,
+        forcing_dirs=forcing_dirs
+    )
+    cy.add(sim)
+    cy.add(job_restart)
     # Parallel test
     cy_parallel = copy.deepcopy(cy)
     cy_dir = pathlib.Path(tmpdir).joinpath('cycle_parallel_run')
@@ -1115,6 +1133,28 @@ def test_cycle_run(
         "Some parallel cycle casts did not run successfully."
     assert cy_dir.joinpath("WrfHydroCycle.pkl").exists()
 
+
+def test_cycle_run_parallel_casts_in_memory(
+    simulation_compiled,
+    job_restart,
+    scheduler,
+    tmpdir,
+    capfd,
+    init_times,
+    restart_dirs
+):
+    sim = simulation_compiled
+    tmp = pathlib.Path(tmpdir)
+    forcing_dirs = [tmp / letter for letter in list(string.ascii_lowercase)[0:len(init_times)]]
+    for dir in forcing_dirs:
+        dir.mkdir()
+    cy = CycleSimulation(
+        init_times=init_times,
+        restart_dirs=restart_dirs,
+        forcing_dirs=forcing_dirs
+    )
+    cy.add(sim)
+    cy.add(job_restart)
     # Parallel test with ensemble in memory
     cy_parallel = copy.deepcopy(cy)
     cy_dir = pathlib.Path(tmpdir).joinpath('cy_parallel_run_cy_in_memory')
@@ -1127,6 +1167,28 @@ def test_cycle_run(
         "Some parallel cycle casts in memory did not run successfully."
     assert cy_dir.joinpath("WrfHydroCycle.pkl").exists()
 
+
+def test_cycle_run_parallel_teams(
+    simulation_compiled,
+    job_restart,
+    scheduler,
+    tmpdir,
+    capfd,
+    init_times,
+    restart_dirs
+):
+    sim = simulation_compiled
+    tmp = pathlib.Path(tmpdir)
+    forcing_dirs = [tmp / letter for letter in list(string.ascii_lowercase)[0:len(init_times)]]
+    for dir in forcing_dirs:
+        dir.mkdir()
+    cy = CycleSimulation(
+        init_times=init_times,
+        restart_dirs=restart_dirs,
+        forcing_dirs=forcing_dirs
+    )
+    cy.add(sim)
+    cy.add(job_restart)
     # Parallel teams test
     cy_teams = copy.deepcopy(cy)
     cy_dir = pathlib.Path(tmpdir).joinpath('cy_team')
